requirements.txt
django==3.1.2
djangorestframework==3.12.1
django-environ==0.4.5
django-redis==4.12.1
gunicorn==20.0.4
boto3==1.16.4
django-storages==1.10.1
Pillow==8.0.1
requests==2.24.0
django-minio-storage==0.3.8
django-stdimage==5.1.1
django-amazon-ses==4.0.0
django-tinymce==3.0.2
mysqlclient==2.0.1

